シフトのマイグレーションファイルの外部キー編集

+++ db/migrate/20241220012916_add_prefecture_id_to_groups.rb
@@ -1,6 +1,7 @@
 class AddPrefectureIdToGroups < ActiveRecord::Migration[6.1]
   def change
-    add_column :groups, :prefecture_id, :integer, null: false
+    # prefecture_idカラムの型をbigintに変更
+    add_column :groups, :prefecture_id, :bigint, null: false
     add_foreign_key :groups, :prefectures
   end
 end
